Fix a* and A* template
The "a*" template would error when unpacking because values[1:byte]
would fail, since bytes = -1. bytes is now set to length(values).

The "A*" template would error when packing because nchar(value) > byte.
byte is now set to nchar(value). The "A*" template would error when
unpacking for the same reason as "a*".