Test pack() on uint8 outside of [0,255] fails
Numbers smaller than 0 and larger than 255 cannot be represented in an
unsigned 8-bit int. Add tests to ensure that pack() fails in this case.

Unsure how to test unpack() though. You can always unpack a byte as an
unsigned 8-bit int, even though it would be incorrect if that wasn't
how it was packed.